feat(quiick load): node adapters refactor
