+++ src-tauri/binaries_versions_esmeralda.json
