+++ src-tauri/src/setup/phase_node.rs
@@ -24,7 +24,7 @@ use std::time::Duration;
 
 use crate::{
     configs::{config_core::ConfigCore, trait_config::ConfigImpl},
-    node_manager::{NodeManagerError, STOP_ON_ERROR_CODES},
+    node::node_manager::{NodeManagerError, STOP_ON_ERROR_CODES},
     progress_trackers::{
         progress_plans::{ProgressPlans, ProgressSetupNodePlan},
         progress_stepper::ProgressStepperBuilder,
